Handle textDocument/superTypes reqs using the index
Handling of textDocument/superTypes now mimics the handling of
textDcoument/subTypes, but searches the index's relations using
RelationKind::DerivedFrom instead of RelationKind::Baseof

Previously, a textDocument/superTypes request was honoured by
searching a TypeHierarchyItem's precomputed state, data.parents, after
a (possibly expensive) serialization round trip to the client.

More seriously, this scheme prevented accurate responses to
textDocument/superTypes and textDocument/subTypes, since the latter
filled in the data.parents field in preparation of the former but
without regard to multiple inheritance.

To make this work, the construction of the in-memory index (not
persisted) is enhanced to also store the reverse relation
RelationKind::DerivedFrom for every RelationKind::BaseOf relation.

The implementation of clang::clangd::fillSuperTypes is simplified.  A
new helper function fillSubOrSuperTypes is introduced to reuse some
code.  Other small code reuse strategies could be considered.

This commit breaks some tests that relied on the details of the
previous implementation (inherited from a pre-LSP 3.17 implementation
of this feature).  The tests are adjusted in a later commit.